requirements.txt
pyspark>=3.4.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
delta-spark>=2.4.0
python-dotenv>=0.19.0
pyyaml>=6.0 